Squash migrations, and remove unused
After the big release of december 2018 a lot of database objects where
removed. They lived on in migrations. All unused database objects have
been removed from migrations, and remaining migrations where squashed.
